+++ src/main/java/com/nineties/bhr/badge/controller/EmpBadgeController.java
@@ -2,6 +2,7 @@ package com.nineties.bhr.badge.controller;
 
 import com.nineties.bhr.badge.domain.EmpBadge;
 import com.nineties.bhr.badge.service.BadgeService;
+import com.nineties.bhr.badge.service.EmpBadgeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -11,20 +12,19 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-@RequestMapping("/myBadge")
+@RequestMapping("/emp")
 public class EmpBadgeController {
 
-    private final BadgeService badgeService;
+    private final EmpBadgeService empBadgeService;
 
     @Autowired
-    public EmpBadgeController(BadgeService badgeService) {
-        this.badgeService = badgeService;
+    public EmpBadgeController(EmpBadgeService empBadgeService) {
+        this.empBadgeService = empBadgeService;
     }
 
-    @GetMapping("/list")
-    public ResponseEntity<List<EmpBadge>> getRecentBadgesForCurrentUser() {
-        List<EmpBadge> recentBadges = badgeService.getRecentBadgesForCurrentUser();
-        return ResponseEntity.ok(recentBadges);
+    @GetMapping("/badge")
+    public List<EmpBadge> getRecentBadgesForCurrentUser() {
+        return empBadgeService.getRecentBadgesForCurrentUser();
     }
 }
