+++ src/main/java/com/nineties/bhr/login/controller/LoginController.java
@@ -48,6 +48,4 @@ public class LoginController {
     }
 
 
-
-
 }
